refactor: 최종 승패 판단 계산하는 객체 생성, 로직 이동
